style(rust): drop unnecessary same-type casts (angr-00pz)
Removes 16 clippy unnecessary_cast warnings (u32->u32, u128->u128)
across 5 files:

- callbacks.rs (3): `(size * 8) as u32` where size is already u32
- claripy_bridge.rs (1): `concrete_val as u128` (already u128)
- exploration/stepping.rs (2): pointer_size/max_steps_per_run both u32
- procedures/fileops.rs (4): pipe() returns (u32, u32)
- vex/ops.rs (6): loop indices `i` and `half_count + i` both u32

Pure syntactic — no behavior change. Each site verified for
narrowing/widening before stripping. cargo check clean; 683/683
rust_exploration tests pass. Clippy backlog: 360 -> 344.

+++ native/angr/src/callbacks.rs
@@ -1305,7 +1305,7 @@ impl PythonCallbacks {
 
             // The callback should return bytes
             let bytes: Vec<u8> = result.extract(py)?;
-            let width = (size * 8) as u32;
+            let width = size * 8;
             let mut value: u128 = 0;
             for (i, &byte) in bytes.iter().enumerate() {
                 if (i * 8) as u32 >= width {
@@ -1319,7 +1319,7 @@ impl PythonCallbacks {
         // Fallback: load from first address only
         if let Some(first_addr) = addrs.first() {
             let (data, _is_symbolic, _ast) = self.call_memory_load(py, *first_addr, size)?;
-            let width = (size * 8) as u32;
+            let width = size * 8;
             let mut value: u128 = 0;
             for (i, &byte) in data.iter().enumerate() {
                 if (i * 8) as u32 >= width {
@@ -1330,7 +1330,7 @@ impl PythonCallbacks {
             Ok(RustBV::concrete(value, width))
         } else {
             // No addresses - return zero
-            Ok(RustBV::zero((size * 8) as u32))
+            Ok(RustBV::zero(size * 8))
         }
     }
